[REF] externalise zpl print into pos_self_service_print_zpl
