+++ pos_self_service_base/readme/USAGE.rst
@@ -12,7 +12,3 @@ Launch Firefox
 This self-service module can be used in kiosk mode. To launch Firefox in kiosk mode, run this command::
 
     firefox <url> -foreground --kiosk
-
-Reset a Zebra printer
-~~~~~~~~~~~~~~~~~~~~~
-`Factory reset video <https://www.youtube.com/watch?v=SXvoW3sm6ls>`_
