+++ pos_self_service_base/views/pos_config_view.xml
@@ -14,33 +14,10 @@
                         </div>
                         <div class="o_setting_right_pane">
                             <label for="iface_self_service"/>
-                            <div class="text-muted">
+                            <div class="text-muted" id="iface_self_service_text">
                             Use your PoS as Self-Service
                             </div>
-                            <div class="content-group" attrs="{'invisible' : [('iface_self_service', '=', False)]}">
-                                <div class="row mt16">
-                                    <label for="printer_name" class="col-lg-3 o_light_label"/>
-                                    <field name="printer_name"/>
-                                </div>
-                                <div class="row">
-                                    <label for="label_offset_x" class="col-lg-3 o_light_label"/>
-                                    <field name="label_offset_x"/>
-                                </div>
-                                <div class="row">
-                                    <label for="label_offset_y" class="col-lg-3 o_light_label"/>
-                                    <field name="label_offset_y"/>
-                                </div>
-                                <div class="row">
-                                    <label for="label_width" class="col-lg-3 o_light_label"/>
-                                    <field name="label_width"/>
-                                </div>
-                                <div class="row">
-                                    <label for="label_height" class="col-lg-3 o_light_label"/>
-                                    <field name="label_height"/>
-                                </div>
-                            </div>
                         </div>
-
                     </div>
                 </div>
             </xpath>
