+++ pos_self_service_print_zpl/__init__.py
@@ -1,0 +1,1 @@
+from . import models
