+++ pos_self_service_print_zpl/models/__init__.py
@@ -1,0 +1,1 @@
+from . import pos_config
