+++ pos_self_service_print_zpl/readme/CONTRIBUTORS.rst
@@ -1,0 +1,5 @@
+* `Coop IT Easy SCRLfs <https://coopiteasy.be>`_:
+    * Grégoire Leeuwerck <gregoire@coopiteasy.be>
+    * Vincent Van Rossem <vincent@coopiteasy.be>
+* `Le Nid <https://lenid.ch>`_:
+    * François Kawala
