+++ pos_self_service_print_zpl/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module handles barcode printing from the browser, using ZPL.
