+++ pos_self_service_print_zpl/readme/USAGE.rst
@@ -1,0 +1,3 @@
+Reset a Zebra printer
+~~~~~~~~~~~~~~~~~~~~~
+`Factory reset video <https://www.youtube.com/watch?v=SXvoW3sm6ls>`_
