+++ pos_self_service_print_zpl/static/src/lib/ipp/out.json
